# waspadai-product-docs/contracts/future/product-api.openapi.yaml
...
components:
  schemas:
    AIResult:
      $ref: '#/components/schemas/VerificationResponse'
...
    HistoryDetail:
      additionalProperties: false
      properties:
        history:
          $ref: '#/components/schemas/HistoryMeta'
        image_url:
          anyOf:
          - format: uri
            type: string
          - type: 'null'
        request_id:
          format: uuid
          type: string
        result:
          $ref: '#/components/schemas/AIResult'
        status:
          enum:
          - COMPLETED
          type: string
        execution_mode:
          type: string
          enum:
          - REMOTE
          - MOCK
          description: Product execution provenance; synthetic fixture has MOCK even
            though strict upstream result.mode is LIVE.
      required:
      - request_id
      - status
      - history
      - result
      - execution_mode
      type: object
...
    HistoryPage:
      additionalProperties: false
      properties:
        items:
          items:
            $ref: '#/components/schemas/HistoryItem'
          type: array
        next_cursor:
          anyOf:
          - type: string
          - type: 'null'
      required:
      - items
      - next_cursor
      type: object
...
    ModuleDetail:
      additionalProperties: false
      properties:
        lessons:
          items:
            $ref: '#/components/schemas/Lesson'
          type: array
        module_id:
          format: uuid
          type: string
        progress_percent:
          maximum: 100
          minimum: 0
          type: number
        slug:
          type: string
        summary:
          type: string
        title:
          type: string
        version:
          minimum: 1
          type: integer
      required:
      - module_id
      - slug
      - title
      - summary
      - version
      - progress_percent
      - lessons
      type: object
    ModulePage:
      additionalProperties: false
      properties:
        items:
          items:
            $ref: '#/components/schemas/ModuleSummary'
          type: array
        next_cursor:
          anyOf:
          - type: string
          - type: 'null'
      required:
      - items
      - next_cursor
      type: object
...
    Progress:
      additionalProperties: false
      properties:
        best_score:
          anyOf:
          - maximum: 100
            minimum: 0
            type: number
          - type: 'null'
        completed_lessons:
          minimum: 0
          type: integer
        latest_score:
          anyOf:
          - maximum: 100
            minimum: 0
            type: number
          - type: 'null'
        module_id:
          format: uuid
          type: string
        progress_percent:
          maximum: 100
          minimum: 0
          type: number
        total_lessons:
          minimum: 0
          type: integer
        updated_at:
          format: date-time
          type: string
      required:
      - module_id
      - completed_lessons
      - total_lessons
      - progress_percent
      - latest_score
      - best_score
      - updated_at
      type: object
...
  securitySchemes:
    SupabaseBearer:
      bearerFormat: JWT
      description: Validated Supabase access token; no internal AI key on clients.
      scheme: bearer
      type: http
  responses:
    SafeError:
      description: Safe Product error envelope; code/action defined in integration
        guide.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
info:
  description: Product target contract aligned to user-uploaded WaspadAI AI OpenAPI
    0.7.0. All upstream schemas copied exactly. Product workflows remain draft; GitHub
    commit and live deployment not verified.
  title: WaspadAI Product API — DRAFT
  version: 3.0.1-draft
openapi: 3.1.0
paths:
...
  /api/v1/history:
    get:
      operationId: get_api_v1_history
      parameters:
      - in: query
        name: limit
        schema:
          default: 20
          maximum: 100
          minimum: 1
          type: integer
      - in: query
        name: cursor
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HistoryPage'
          description: Successful target response
        '400':
          $ref: '#/components/responses/SafeError'
        '401':
          $ref: '#/components/responses/SafeError'
        '403':
          $ref: '#/components/responses/SafeError'
        '404':
          $ref: '#/components/responses/SafeError'
        '409':
          $ref: '#/components/responses/SafeError'
        '413':
          $ref: '#/components/responses/SafeError'
        '415':
          $ref: '#/components/responses/SafeError'
        '422':
          $ref: '#/components/responses/SafeError'
        '429':
          $ref: '#/components/responses/SafeError'
        '502':
          $ref: '#/components/responses/SafeError'
        '503':
          $ref: '#/components/responses/SafeError'
        '504':
          $ref: '#/components/responses/SafeError'
      security:
      - SupabaseBearer: []
      summary: History GET
      tags:
      - History
      x-implementation-status: TARGET_DRAFT
  /api/v1/history/{case_id}:
    delete:
      operationId: delete_api_v1_history_case_id
      parameters:
      - in: path
        name: case_id
        required: true
        schema:
          format: uuid
          type: string
      responses:
        '204':
          description: Completed; no response body
        '400':
          $ref: '#/components/responses/SafeError'
        '401':
          $ref: '#/components/responses/SafeError'
        '403':
          $ref: '#/components/responses/SafeError'
        '404':
          $ref: '#/components/responses/SafeError'
        '409':
          $ref: '#/components/responses/SafeError'
        '413':
          $ref: '#/components/responses/SafeError'
        '415':
          $ref: '#/components/responses/SafeError'
        '422':
          $ref: '#/components/responses/SafeError'
        '429':
          $ref: '#/components/responses/SafeError'
        '502':
          $ref: '#/components/responses/SafeError'
        '503':
          $ref: '#/components/responses/SafeError'
        '504':
          $ref: '#/components/responses/SafeError'
      security:
      - SupabaseBearer: []
      summary: History DELETE
      tags:
      - History
      x-implementation-status: TARGET_DRAFT
    get:
      operationId: get_api_v1_history_case_id
      parameters:
      - in: path
        name: case_id
        required: true
        schema:
          format: uuid
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HistoryDetail'
          description: Successful target response
        '400':
          $ref: '#/components/responses/SafeError'
        '401':
          $ref: '#/components/responses/SafeError'
        '403':
          $ref: '#/components/responses/SafeError'
        '404':
          $ref: '#/components/responses/SafeError'
        '409':
          $ref: '#/components/responses/SafeError'
        '413':
          $ref: '#/components/responses/SafeError'
        '415':
          $ref: '#/components/responses/SafeError'
        '422':
          $ref: '#/components/responses/SafeError'
        '429':
          $ref: '#/components/responses/SafeError'
        '502':
          $ref: '#/components/responses/SafeError'
        '503':
          $ref: '#/components/responses/SafeError'
        '504':
          $ref: '#/components/responses/SafeError'
      security:
      - SupabaseBearer: []
      summary: History GET
      tags:
      - History
      x-implementation-status: TARGET_DRAFT
...
  /api/v1/learning/lessons/{lesson_id}/complete:
    post:
      operationId: post_api_v1_learning_lessons_lesson_id_complete
      parameters:
      - in: path
        name: lesson_id
        required: true
        schema:
          format: uuid
          type: string
      - description: Product-local idempotency, not a guarantee of upstream exactly-once
          inference.
        in: header
        name: Idempotency-Key
        required: true
        schema:
          format: uuid
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Progress'
          description: Successful target response
        '400':
          $ref: '#/components/responses/SafeError'
        '401':
          $ref: '#/components/responses/SafeError'
        '403':
          $ref: '#/components/responses/SafeError'
        '404':
          $ref: '#/components/responses/SafeError'
        '409':
          $ref: '#/components/responses/SafeError'
        '413':
          $ref: '#/components/responses/SafeError'
        '415':
          $ref: '#/components/responses/SafeError'
        '422':
          $ref: '#/components/responses/SafeError'
        '429':
          $ref: '#/components/responses/SafeError'
        '502':
          $ref: '#/components/responses/SafeError'
        '503':
          $ref: '#/components/responses/SafeError'
        '504':
          $ref: '#/components/responses/SafeError'
      security:
      - SupabaseBearer: []
      summary: Learning POST
      tags:
      - Learning
      x-implementation-status: TARGET_DRAFT
  /api/v1/learning/modules:
    get:
      operationId: get_api_v1_learning_modules
      parameters:
      - in: query
        name: limit
        schema:
          default: 20
          maximum: 100
          minimum: 1
          type: integer
      - in: query
        name: cursor
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModulePage'
          description: Successful target response
        '400':
          $ref: '#/components/responses/SafeError'
        '401':
          $ref: '#/components/responses/SafeError'
        '403':
          $ref: '#/components/responses/SafeError'
        '404':
          $ref: '#/components/responses/SafeError'
        '409':
          $ref: '#/components/responses/SafeError'
        '413':
          $ref: '#/components/responses/SafeError'
        '415':
          $ref: '#/components/responses/SafeError'
        '422':
          $ref: '#/components/responses/SafeError'
        '429':
          $ref: '#/components/responses/SafeError'
        '502':
          $ref: '#/components/responses/SafeError'
        '503':
          $ref: '#/components/responses/SafeError'
        '504':
          $ref: '#/components/responses/SafeError'
      security:
      - SupabaseBearer: []
      summary: Learning GET
      tags:
      - Learning
      x-implementation-status: TARGET_DRAFT
  /api/v1/learning/modules/{module_id}:
    get:
      operationId: get_api_v1_learning_modules_module_id
      parameters:
      - in: path
        name: module_id
        required: true
        schema:
          format: uuid
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModuleDetail'
          description: Successful target response
        '400':
          $ref: '#/components/responses/SafeError'
        '401':
          $ref: '#/components/responses/SafeError'
        '403':
          $ref: '#/components/responses/SafeError'
        '404':
          $ref: '#/components/responses/SafeError'
        '409':
          $ref: '#/components/responses/SafeError'
        '413':
          $ref: '#/components/responses/SafeError'
        '415':
          $ref: '#/components/responses/SafeError'
        '422':
          $ref: '#/components/responses/SafeError'
        '429':
          $ref: '#/components/responses/SafeError'
        '502':
          $ref: '#/components/responses/SafeError'
        '503':
          $ref: '#/components/responses/SafeError'
        '504':
          $ref: '#/components/responses/SafeError'
      security:
      - SupabaseBearer: []
      summary: Learning GET
      tags:
      - Learning
      x-implementation-status: TARGET_DRAFT
  /api/v1/learning/modules/{module_id}/quiz:
    get:
      operationId: get_api_v1_learning_modules_module_id_quiz
      parameters:
      - in: path
        name: module_id
        required: true
        schema:
          format: uuid
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Quiz'
          description: Successful target response
        '400':
          $ref: '#/components/responses/SafeError'
        '401':
          $ref: '#/components/responses/SafeError'
        '403':
          $ref: '#/components/responses/SafeError'
        '404':
          $ref: '#/components/responses/SafeError'
        '409':
          $ref: '#/components/responses/SafeError'
        '413':
          $ref: '#/components/responses/SafeError'
        '415':
          $ref: '#/components/responses/SafeError'
        '422':
          $ref: '#/components/responses/SafeError'
        '429':
          $ref: '#/components/responses/SafeError'
        '502':
          $ref: '#/components/responses/SafeError'
        '503':
          $ref: '#/components/responses/SafeError'
        '504':
          $ref: '#/components/responses/SafeError'
      security:
      - SupabaseBearer: []
      summary: Learning GET
      tags:
      - Learning
      x-implementation-status: TARGET_DRAFT
  /api/v1/learning/modules/{module_id}/quiz-attempts:
    post:
      operationId: post_api_v1_learning_modules_module_id_quiz_attempts
      parameters:
      - in: path
        name: module_id
        required: true
        schema:
          format: uuid
          type: string
      - description: Product-local idempotency, not a guarantee of upstream exactly-once
          inference.
        in: header
        name: Idempotency-Key
        required: true
        schema:
          format: uuid
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuizRequest'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuizResult'
          description: Successful target response
        '400':
          $ref: '#/components/responses/SafeError'
        '401':
          $ref: '#/components/responses/SafeError'
        '403':
          $ref: '#/components/responses/SafeError'
        '404':
          $ref: '#/components/responses/SafeError'
        '409':
          $ref: '#/components/responses/SafeError'
        '413':
          $ref: '#/components/responses/SafeError'
        '415':
          $ref: '#/components/responses/SafeError'
        '422':
          $ref: '#/components/responses/SafeError'
        '429':
          $ref: '#/components/responses/SafeError'
        '502':
          $ref: '#/components/responses/SafeError'
        '503':
          $ref: '#/components/responses/SafeError'
        '504':
          $ref: '#/components/responses/SafeError'
      security:
      - SupabaseBearer: []
      summary: Learning POST
      tags:
      - Learning
      x-implementation-status: TARGET_DRAFT
  /api/v1/learning/progress:
    get:
      operationId: get_api_v1_learning_progress
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProgressPage'
          description: Successful target response
        '400':
          $ref: '#/components/responses/SafeError'
        '401':
          $ref: '#/components/responses/SafeError'
        '403':
          $ref: '#/components/responses/SafeError'
        '404':
          $ref: '#/components/responses/SafeError'
        '409':
          $ref: '#/components/responses/SafeError'
        '413':
          $ref: '#/components/responses/SafeError'
        '415':
          $ref: '#/components/responses/SafeError'
        '422':
          $ref: '#/components/responses/SafeError'
        '429':
          $ref: '#/components/responses/SafeError'
        '502':
          $ref: '#/components/responses/SafeError'
        '503':
          $ref: '#/components/responses/SafeError'
        '504':
          $ref: '#/components/responses/SafeError'
      security:
      - SupabaseBearer: []
      summary: Learning GET
      tags:
      - Learning
      x-implementation-status: TARGET_DRAFT
...
  /api/v1/verifications/image:
    post:
      operationId: post_api_v1_verifications_image
      parameters:
      - description: Product-local idempotency, not a guarantee of upstream exactly-once
          inference.
        in: header
        name: Idempotency-Key
        required: true
        schema:
          format: uuid
          type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ImageRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerificationEnvelope'
          description: Successful target response
        '400':
          $ref: '#/components/responses/SafeError'
        '401':
          $ref: '#/components/responses/SafeError'
        '403':
          $ref: '#/components/responses/SafeError'
        '404':
          $ref: '#/components/responses/SafeError'
        '409':
          $ref: '#/components/responses/SafeError'
        '413':
          $ref: '#/components/responses/SafeError'
        '415':
          $ref: '#/components/responses/SafeError'
        '422':
          $ref: '#/components/responses/SafeError'
        '429':
          $ref: '#/components/responses/SafeError'
        '502':
          $ref: '#/components/responses/SafeError'
        '503':
          $ref: '#/components/responses/SafeError'
        '504':
          $ref: '#/components/responses/SafeError'
      security:
      - SupabaseBearer: []
      summary: Verification POST
      tags:
      - Verification
      x-implementation-status: TARGET_DRAFT
  /api/v1/verifications/text:
    post:
      operationId: post_api_v1_verifications_text
      parameters:
      - description: Product-local idempotency, not a guarantee of upstream exactly-once
          inference.
        in: header
        name: Idempotency-Key
        required: true
        schema:
          format: uuid
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TextRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerificationEnvelope'
          description: Successful target response
        '400':
          $ref: '#/components/responses/SafeError'
        '401':
          $ref: '#/components/responses/SafeError'
        '403':
          $ref: '#/components/responses/SafeError'
        '404':
          $ref: '#/components/responses/SafeError'
        '409':
          $ref: '#/components/responses/SafeError'
        '413':
          $ref: '#/components/responses/SafeError'
        '415':
          $ref: '#/components/responses/SafeError'
        '422':
          $ref: '#/components/responses/SafeError'
        '429':
          $ref: '#/components/responses/SafeError'
        '502':
          $ref: '#/components/responses/SafeError'
        '503':
          $ref: '#/components/responses/SafeError'
        '504':
          $ref: '#/components/responses/SafeError'
      security:
      - SupabaseBearer: []
      summary: Verification POST
      tags:
      - Verification
      x-implementation-status: TARGET_DRAFT
servers:
- description: Local Product target only
  url: http://127.0.0.1:8001
...